Move spritesheet to canvas folder
